requirements-dev.txt
-r requirements.txt
pytest==8.1.1
pytest-flask==1.3.0
